requirements.txt
python-dotenv>=1.0.0
openai>=1.0.0
streamlit>=1.28.0
pathlib2>=2.3.7
typing-extensions>=4.0.0
